requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==26.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
datasets==4.8.4
dill==0.4.1
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docker==7.1.0
fastapi==0.133.1
filelock==3.25.2
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2026.2.0
gevent==25.9.1
greenlet==3.3.1
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.7
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.19
numpy==2.3.5
packaging==25.0
pandas==3.0.1
pillow==12.0.0
propcache==0.4.1
psycopg2==2.9.11
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.3.0
requests==2.32.5
rich==14.3.3
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
starlette==0.52.1
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websocket==0.2.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zope.event==6.1
zope.interface==8.2
